Falta probar la función crearRegistros se ha quedado colagada el virtualbox
