+++ ejercicio1/src/main/java/com/team2/u22/ejercicio1/controlador/ClienteControler.java
@@ -3,6 +3,9 @@ package com.team2.u22.ejercicio1.controlador;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+import com.team2.u22.ejercicio1.modelo.Cliente;
+import com.team2.u22.ejercicio1.vista.Vista;
+
 /**
  * 
  * ejercicio1 - com.team2.u22.ejercicio1.controlador - ClienteControler
@@ -16,16 +19,17 @@ import java.awt.event.ActionListener;
 public class ClienteControler implements ActionListener{
 	
 	private Cliente cliente;
-	private ClienteVista clienteVista;
+	private Vista vista;
 
 	/**
 	 * Contructor de clase le pasamos el moledo cliente y vista cliente
 	 * @param cliente
 	 * @param clienteVista
 	 */
-	public ClienteControler(Cliente cliente, ClienteVista clienteVista) {
+	public ClienteControler(Cliente cliente, Vista clienteVista) {
 		this.cliente = cliente;
-		this.clienteVista = clienteVista;
+		this.vista = clienteVista;
+		//this.vista.bot
 	}
