+++ ejercicio1/src/main/java/com/team2/u22/ejercicio1/vista/Vista.java
@@ -40,22 +40,6 @@ public class Vista extends JFrame {
 	private JTextField idBUTextField;
 
 	/**
-	 * Launch the application.
-	 */
-	public static void main(String[] args) {
-		EventQueue.invokeLater(new Runnable() {
-			public void run() {
-				try {
-					Vista frame = new Vista();
-					frame.setVisible(true);
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-			}
-		});
-	}
-
-	/**
 	 * Create the frame.
 	 */
 	public Vista() {
